requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
oauthlib==2.1.0
PyMySQL==0.8.1
PySocks==1.6.8
python-decouple==3.1
requests==2.20.0
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.6.0
urllib3==1.22